refactor: receipt loading, download block
