+++ frontend/src/features/payment/stripe/components/StripePaymentBlock.tsx
@@ -21,16 +21,14 @@ import { FormColorTheme, FormResponseMode } from '~shared/types/form'
 import { centsToDollars } from '~utils/payments'
 import Button from '~components/Button'
 
+import { FormPaymentPageProps } from '~features/payment/FormPaymentPage'
 import { usePublicFormContext } from '~features/public-form/PublicFormContext'
-
-import { FormPaymentPageProps } from '../FormPaymentPage'
 
 // Make sure to call `loadStripe` outside of a component’s render to avoid
 // recreating the `Stripe` object on every render.
 
 export interface PaymentPageBlockProps extends FormPaymentPageProps {
   focusOnMount?: boolean
-  // TODO: want to refactor FormPaymentPageProps into payment
   triggerPaymentStatusRefetch: () => void
 }
 
@@ -124,7 +122,7 @@ const StripeCheckoutForm = ({
   )
 }
 
-export const StripePaymentModal = ({
+export const StripePaymentBlock = ({
   submissionId,
   paymentClientSecret,
   publishableKey,
